wip, lib not linking
