+++ src/bsp/include/bsp.h
@@ -1,0 +1,10 @@
+#ifndef __BSP_H_
+#define __BSP_H_
+
+#define BSP_TEST 4;
+
+void init_chip();
+
+void blinky();
+
+#endif // __BSP_H_
